Split package:http_profile into multiple files
