+++ pkgs/http_profile/lib/src/http_profile.dart
@@ -1,0 +1,14 @@
+// Copyright (c) 2024, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'dart:async';
+import 'dart:developer' show addHttpClientProfilingData, Service;
+import 'dart:io' show HttpClient, HttpClientResponseCompressionState;
+import 'dart:isolate' show Isolate;
+
+import 'utils.dart';
+
+part 'http_client_request_profile.dart';
+part 'http_profile_request_data.dart';
+part 'http_profile_response_data.dart';
